Works on 64bit supporting.
Fixed calling of MAP_EVENT_OUT_CRITTER, MAP_EVENT_IN_CRITTER.
Fixed global group sending.

git-svn-id: https://cvet.by:8443/svn/FOnlineSource@187 c2c6a35e-ad2b-4e1f-bb64-97b989be360b

Former-commit-id: 29287e21d6c13a1089bad3c8af67bf1b06316fcf

+++ Source/Critter.h
@@ -218,7 +218,7 @@ public:
 	uint CountItemPid(ushort item_pid);
 	void TakeDefaultItem(uchar slot);
 	bool MoveItem(uchar from_slot, uchar to_slot, uint item_id, uint count);
-	uint RealCountItems(){return invItems.size();}
+	uint RealCountItems(){return (uint)invItems.size();}
 	uint CountItems();
 	ItemPtrVec& GetInventory(){SyncLockItems(); return invItems;}
 	bool IsHaveGeckItem();
